+++ src/main/java/de/muenchen/zammad/ldap/config/RoleIds.java
@@ -1,9 +1,9 @@
-package de.muenchen.zammad.ldap;
+package de.muenchen.zammad.ldap.config;
 
 import org.springframework.stereotype.Component;
 
+import de.muenchen.zammad.ldap.property.ZammadProperties;
 import de.muenchen.zammad.ldap.service.ZammadService;
-import de.muenchen.zammad.ldap.service.config.ZammadProperties;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
