+++ src/main/java/de/muenchen/zammad/ldap/service/ZammadSyncServiceSubtree.java
@@ -15,8 +15,8 @@ import de.muenchen.oss.ezldap.core.LdapUserDTO;
 import de.muenchen.zammad.ldap.domain.Signatures;
 import de.muenchen.zammad.ldap.domain.ZammadGroupDTO;
 import de.muenchen.zammad.ldap.domain.ZammadUserDTO;
-import de.muenchen.zammad.ldap.service.config.OrganizationalUnitsCommonProperties;
-import de.muenchen.zammad.ldap.service.config.ZammadProperties;
+import de.muenchen.zammad.ldap.property.OrganizationalUnitsCommonProperties;
+import de.muenchen.zammad.ldap.property.ZammadProperties;
 import de.muenchen.zammad.ldap.tree.LdapOuNode;
 import lombok.Getter;
 import lombok.Setter;
