+++ src/test/java/de/muenchen/zammad/ldap/service/FindSignatureTest.java
@@ -17,7 +17,7 @@ import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 
 import de.muenchen.zammad.ldap.domain.Signatures;
-import de.muenchen.zammad.ldap.service.config.OrganizationalUnitsCommonProperties;
+import de.muenchen.zammad.ldap.property.OrganizationalUnitsCommonProperties;
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
